requirements.txt
numpy
pillow
scikit-learn
torch
torchvision
tqdm
transformers
